requirements.txt
openai>=1.40.0
playwright>=1.48.0
httpx>=0.27.0
python-dotenv>=1.0.0
